QuPhys

Documentation for QuPhys.

QuPhys.mcsolve

QuPhys.mesolve

QuPhys.sesolve

QuPhys.mcsolve — Method

function mcsolve(H::QuantumObject{<:AbstractArray{T}, OperatorQuantumObject}, 
        ψ0::QuantumObject{<:AbstractArray{T}, KetQuantumObject}, 
        t_l::AbstractVector, c_ops::AbstractVector;
        e_ops::AbstractVector = [], 
        n_traj::Int = 1,
        batch_size::Int = min(10, n_traj),
        alg = AutoVern7(KenCarp4(autodiff=false)),
        ensemble_method = EnsembleThreads(), 
        H_t = nothing,
        progress = true,
        kwargs...)

Time evolution of an open quantum system using quantum trajectories.

QuPhys.mesolve — Method

function mesolve(H::QuantumObject{<:AbstractArray{T}, HOpType}, 
        ψ0::QuantumObject{<:AbstractArray{T}, StateOpType},
        t_l::AbstractVector, c_ops::AbstractVector; 
        e_ops::AbstractVector = [], 
        alg = LinearExponential(krylov=:simple), 
        H_t = nothing, 
        progress = true,
        kwargs...)

Time evolution of an open quantum system using master equation.

QuPhys.sesolve — Method

function sesolve(H::QuantumObject{<:AbstractArray{T}, OperatorQuantumObject},
            ψ0::QuantumObject{<:AbstractArray{T}, KetQuantumObject},
            t_l::AbstractVector;  
            e_ops::AbstractVector = [], 
            alg = LinearExponential(), 
            H_t = nothing, 
            progress = true,
            kwargs...)

Time evolution of a closed quantum system using Schrödinger equation.