API

General functions

QuPhys.row_major_reshape — Function

row_major_reshape(Q::AbstractArray, shapes...)

Reshapes Q in the row-major order, as numpy.

QuPhys.meshgrid — Function

meshgrid(x::AbstractVector, y::AbstractVector)

Equivalent to numpy meshgrid.

QuPhys.gaussian — Function

gaussian(x, μ::Real, σ::Real)

Returns the gaussian function $\exp \left[- \frac{(x - \mu)^2}{2 \sigma^2} \right]$, where $\mu$ and $\sigma^2$ are the mean and the variance respectively.

Time evolution

QuPhys.sesolve — Function

function sesolve(H::QuantumObject{<:AbstractArray{T}, OperatorQuantumObject},
            ψ0::QuantumObject{<:AbstractArray{T}, KetQuantumObject},
            t_l::AbstractVector;  
            e_ops::AbstractVector = [], 
            alg = LinearExponential(), 
            H_t = nothing, 
            params::AbstractVector = [],
            progress::Bool = true,
            callbacks = [],
            kwargs...)

Time evolution of a closed quantum system using Schrödinger equation.

QuPhys.mesolve — Function

function mesolve(H::QuantumObject{<:AbstractArray{T}, HOpType}, 
        ψ0::QuantumObject{<:AbstractArray{T}, StateOpType},
        t_l::AbstractVector, c_ops::AbstractVector; 
        e_ops::AbstractVector = [], 
        alg = LinearExponential(krylov=:simple), 
        H_t = nothing, 
        params::AbstractVector = [],
        progress::Bool = true,
        callbacks = [],
        kwargs...)

Time evolution of an open quantum system using master equation.

QuPhys.mcsolve — Function

function mcsolve(H::QuantumObject{<:AbstractArray{T}, OperatorQuantumObject}, 
        ψ0::QuantumObject{<:AbstractArray{T}, KetQuantumObject}, 
        t_l::AbstractVector, c_ops::AbstractVector;
        e_ops::AbstractVector = [], 
        n_traj::Int = 1,
        batch_size::Int = min(10, n_traj),
        alg = AutoVern7(KenCarp4(autodiff=false)),
        ensemble_method = EnsembleThreads(), 
        H_t = nothing,
        progress::Bool = true,
        jump_interp_pts::Int = 0,
        callbacks = [],
        kwargs...)

Time evolution of an open quantum system using quantum trajectories.

QuPhys.dfd_mesolve — Function

function dfd_mesolve(H::Function, ψ0::QuantumObject{<:AbstractArray{T}, StateOpType},
    t_l::AbstractVector, c_ops::Function, e_ops::Function, maxdims::AbstractVector;
    tol_list::AbstractVector = [],
    alg = Vern7(),
    progress::Bool = true,
    callbacks::AbstractVector = [],
    kwargs...)

Time evolution of an open quantum system using master equation, dynamically changing the dimension of the Hilbert subspaces.