API

Quantum object functions

QuPhys.BraQuantumObject — Type

BraQuantumObject <: QuantumObjectType

Abstract type representing a bra state $\bra{\psi}$.

QuPhys.KetQuantumObject — Type

KetQuantumObject <: QuantumObjectType

Abstract type representing a ket state $\ket{\psi}$.

QuPhys.OperatorQuantumObject — Type

OperatorQuantumObject <: QuantumObjectType

Abstract type representing an operator $\hat{O}$.

QuPhys.SuperOperatorQuantumObject — Type

SuperOperatorQuantumObject <: QuantumObjectType

Abstract type representing a super-operator $\hat{\mathcal{O}}$.

QuPhys.QuantumObject — Type

mutable struct QuantumObject{MT<:AbstractArray,ObjType<:QuantumObjectType}
    data::MT
    type::Type{ObjType}
    dims::Vector{Int}
end

Julia struct representing any quantum operator.

Examples

julia> a = destroy(20)
Quantum Object:   type=Operator   dims=[20]   size=(20, 20)   ishermitian=false
20×20 SparseMatrixCSC{ComplexF64, Int64} with 19 stored entries:
⠈⠢⡀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠈⠢⡀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠈⠢⡀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠈⠢⡀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠈⠢

julia> a isa QuantumObject
true

QuPhys.ket2dm — Function

ket2dm(ψ::QuantumObject)

Transform the ket state $\ket{\psi}$ into a pure density matrix $\hat{\rho} = \dyad{\psi}$.

QuPhys.isbra — Function

isbra(A::QuantumObject)

Checks if the QuantumObject A is a BraQuantumObject state.

QuPhys.isket — Function

isket(A::QuantumObject)

Checks if the QuantumObject A is a KetQuantumObject state.

QuPhys.isoper — Function

isoper(A::QuantumObject)

Checks if the QuantumObject A is a OperatorQuantumObject state.

QuPhys.issuper — Function

issuper(A::QuantumObject)

Checks if the QuantumObject A is a SuperOperatorQuantumObject state.

Base.size — Function

size(A::QuantumObject)

Returns the size of the matrix or vector corresponding to the QuantumObject A.

Base.length — Function

length(A::QuantumObject)

Returns the length of the matrix or vector corresponding to the QuantumObject A.

LinearAlgebra.tr — Function

tr(A::QuantumObject})

Returns the trace of A.

Examples

julia> a = destroy(20)
Quantum Object:   type=Operator   dims=[20]   size=(20, 20)   ishermitian=false
20×20 SparseMatrixCSC{ComplexF64, Int64} with 19 stored entries:
⠈⠢⡀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠈⠢⡀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠈⠢⡀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠈⠢⡀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠈⠢

julia> tr(a' * a)
190.0 + 0.0im

LinearAlgebra.norm — Function

norm(A::QuantumObject)

Returns the norm of A.

Examples

julia> ψ = fock(10, 2)
Quantum Object:   type=Ket   dims=[10]   size=(10,)
10-element Vector{ComplexF64}:
 0.0 + 0.0im
 0.0 + 0.0im
 1.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im

julia> norm(ψ)
1.0

Base.kron — Function

kron(A::QuantumObject, B::QuantumObject)

Returns the Kronecker product $\hat{A} \otimes \hat{B}$.

Examples

julia> a = destroy(20)
Quantum Object:   type=Operator   dims=[20]   size=(20, 20)   ishermitian=false
20×20 SparseMatrixCSC{ComplexF64, Int64} with 19 stored entries:
⠈⠢⡀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠈⠢⡀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠈⠢⡀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠈⠢⡀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠈⠢

julia> kron(a, a)
Quantum Object:   type=Operator   dims=[20, 20]   size=(400, 400)   ishermitian=false
400×400 SparseMatrixCSC{ComplexF64, Int64} with 361 stored entries:
⠀⠀⠘⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢦⡀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⠦

General functions

QuPhys.row_major_reshape — Function

row_major_reshape(Q::AbstractArray, shapes...)

Reshapes Q in the row-major order, as numpy.

QuPhys.meshgrid — Function

meshgrid(x::AbstractVector, y::AbstractVector)

Equivalent to numpy meshgrid.

QuPhys.sparse_to_dense — Function

sparse_to_dense(A::QuantumObject)

Converts a sparse QuantumObject to a dense QuantumObject.

QuPhys.dense_to_sparse — Function

dense_to_sparse(A::QuantumObject)

Converts a dense QuantumObject to a sparse QuantumObject.

QuPhys.tidyup — Function

tidyup(A::QuantumObject, tol::Real=1e-14)

Removes those elements of a QuantumObject A whose absolute value is less than tol.

QuPhys.tidyup! — Function

tidyup!(A::QuantumObject, tol::Real=1e-14)

Removes those elements of a QuantumObject A whose absolute value is less than tol. In-place version of tidyup.

QuPhys.gaussian — Function

gaussian(x::Number, μ::Number, σ::Number)

Returns the gaussian function $\exp \left[- \frac{(x - \mu)^2}{2 \sigma^2} \right]$, where $\mu$ and $\sigma^2$ are the mean and the variance respectively.

QuPhys.ptrace — Function

ptrace(QO::QuantumObject, sel::Vector{Int})

Partial trace of a quantum state QO leaving only the dimensions with the indices present in the sel vector.

Examples

Two qubits in the state $\ket{\psi} = \ket{e,g}$:

julia> ψ = kron(fock(2,0), fock(2,1))
Quantum Object:   type=Ket   dims=[2, 2]   size=(4,)
4-element Vector{ComplexF64}:
 0.0 + 0.0im
 1.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im

julia> ptrace(ψ, [2])
Quantum Object:   type=Operator   dims=[2]   size=(2, 2)   ishermitian=true
2×2 Matrix{ComplexF64}:
 0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im

or in an entangled state $\ket{\psi} = \frac{1}{\sqrt{2}} \left( \ket{e,e} + \ket{g,g} \right)$:

julia> ψ = 1 / √2 * (kron(fock(2,0), fock(2,0)) + kron(fock(2,1), fock(2,1)))
Quantum Object:   type=Ket   dims=[2, 2]   size=(4,)
4-element Vector{ComplexF64}:
 0.7071067811865475 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
 0.7071067811865475 + 0.0im

julia> ptrace(ψ, [1])
Quantum Object:   type=Operator   dims=[2]   size=(2, 2)   ishermitian=true
2×2 Matrix{ComplexF64}:
 0.5+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.0im

QuPhys.entropy_vn — Function

entropy_vn(ρ::QuantumObject; base::Int=0, tol::Real=1e-15)

Calculates the Von Neumann entropy $S = - \Tr \left[ \hat{\rho} \log \left( \hat{\rho} \right) \right]$ where $\hat{\rho}$ is the density matrix of the system.

The base parameter specifies the base of the logarithm to use, and when using the default value 0, the natural logarithm is used. The tol parameter describes the absolute tolerance for detecting the zero-valued eigenvalues of the density matrix $\hat{\rho}$.

Examples

Pure state:

julia> ψ = fock(2,0)
Quantum Object:   type=Ket   dims=[2]   size=(2,)
2-element Vector{ComplexF64}:
 1.0 + 0.0im
 0.0 + 0.0im

julia> ρ = ket2dm(ψ)
Quantum Object:   type=Operator   dims=[2]   size=(2, 2)   ishermitian=true
2×2 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im

julia> entropy_vn(ρ, base=2)
-0.0

Mixed state:

julia> ρ = 1 / 2 * ( ket2dm(fock(2,0)) + ket2dm(fock(2,1)) )
Quantum Object:   type=Operator   dims=[2]   size=(2, 2)   ishermitian=true
2×2 Matrix{ComplexF64}:
 0.5+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.0im

julia> entropy_vn(ρ, base=2)
1.0

QuPhys.entanglement — Function

entanglement(QO::QuantumObject, sel::Vector)

Calculates the entanglement by doing the partial trace of QO, selecting only the dimensions with the indices contained in the sel vector, and then using the Von Neumann entropy entropy_vn.

QuPhys.expect — Function

expect(O::QuantumObject, ψ::QuantumObject)

Expectation value of the operator O with the state ψ. The latter can be a KetQuantumObject, BraQuantumObject or a OperatorQuantumObject. If ψ is a density matrix, the function calculates $\Tr \left[ \hat{O} \hat{\psi} \right]$, while if ψ is a state, the function calculates $\mel{\psi}{\hat{O}}{\psi}$.

The function returns a real number if the operator is hermitian, and returns a complex number otherwise.

Examples

julia> ψ = 1 / √2 * (fock(10,2) + fock(10,4));

julia> a = destroy(10);

julia> expect(a' * a, ψ) ≈ 3
true

QuPhys.wigner — Function

wigner(state::QuantumObject, xvec::AbstractVector, yvec::AbstractVector; g::Real=√2,
    solver::WignerSolver=WignerLaguerre())

Generates the Wigner quasipropability distribution of state at points xvec + 1im * yvec. The g parameter is a scaling factor related to the value of $\hbar$ in the commutation relation $[x, y] = i \hbar$ via $\hbar=2/g^2$ giving the default value $\hbar=1$.

The solver parameter can be either WignerLaguerre() or WignerClenshaw(). The former uses the Laguerre polynomial expansion of the Wigner function, while the latter uses the Clenshaw algorithm. The Laguerre expansion is faster for sparse matrices, while the Clenshaw algorithm is faster for dense matrices. The WignerLaguerre solver has an optional parallel parameter which defaults to true and uses multithreading to speed up the calculation.

QuPhys.get_coherence — Function

get_coherence(ψ::QuantumObject)

Get the coherence value $\alpha$ by measuring the expectation value of the destruction operator $\hat{a}$ on the state $\ket{\psi}$.

It returns both $\alpha$ and the state $\ket{\delta_\psi} = \exp ( \bar{\alpha} \hat{a} - \alpha \hat{a}^\dagger )$. The latter corresponds to the quantum fulctuations around the coherent state $\ket{\alpha}$.

QuPhys.n_th — Function

n_th(ω::Number, T::Real)

Gives the mean number of excitations in a mode with frequency ω at temperature T: $n_{\rm th} (\omega, T) = \frac{1}{e^{\omega/T} - 1}$

QuPhys.get_data — Function

get_data(A::QuantumObject)

Returns the data of a QuantumObject.

QuPhys.mat2vec — Function

mat2vec(A::AbstractMatrix)

Converts a matrix to a vector.

QuPhys.vec2mat — Function

vec2mat(A::AbstractVector)

Converts a vector to a matrix.

Quantum states, operators and super-operators

QuPhys.spre — Function

spre(O::QuantumObject)

Returns the super-operator form of O acting on the left of the density matrix operator, $\mathcal{O} \left(\hat{O}\right) \left[ \hat{\rho} \right] = \hat{O} \hat{\rho}$.

Since the density matrix is vectorized, this super-operator is always a matrix, obtained from $\mathcal{O} \left(\hat{O}\right) \boldsymbol{\cdot} = \hat{\mathbb{1}} \otimes \hat{O}$.

QuPhys.spost — Function

spost(O::QuantumObject)

Returns the super-operator form of O acting on the right of the density matrix operator, $\mathcal{O} \left(\hat{O}\right) \left[ \hat{\rho} \right] = \hat{\rho} \hat{O}$.

Since the density matrix is vectorized, this super-operator is always a matrix, obtained from $\mathcal{O} \left(\hat{O}\right) \boldsymbol{\cdot} = \hat{O}^T \otimes \hat{\mathbb{1}}$.

QuPhys.sprepost — Function

sprepost(A::QuantumObject, B::QuantumObject)

Returns the super-operator form of A and B acting on the left and the right of the density matrix operator respectively, $\mathcal{O} \left( \hat{A}, \hat{B} \right) \left[ \hat{\rho} \right] = \hat{A} \hat{\rho} \hat{B}$.

Since the density matrix is vectorized, this super-operator is always a matrix, obtained from $\mathcal{O} \left(\hat{A}, \hat{B}\right) \boldsymbol{\cdot} = \text{spre}(A) * \text{spost}(B)$.

QuPhys.lindblad_dissipator — Function

lindblad_dissipator(O::QuantumObject)

Returns the Lindblad super-operator defined as $\mathcal{D} \left( \hat{O} \right) \left[ \hat{\rho} \right] = \frac{1}{2} \left( 2 \hat{O} \hat{\rho} \hat{O}^\dagger - \hat{O}^\dagger \hat{O} \hat{\rho} - \hat{\rho} \hat{O}^\dagger \hat{O} \right)$ considering the density matrix $\hat{\rho}$ in the vectorized form.

QuPhys.destroy — Function

destroy(N::Int)

Bosonic annihilation operator with Hilbert space cutoff N. This operator acts on a fock state as $\hat{a} \ket{n} = \sqrt{n} \ket{n-1}$.

Examples

julia> a = destroy(20)
Quantum Object:   type=Operator   dims=[20]   size=(20, 20)   ishermitian=false
20×20 SparseMatrixCSC{ComplexF64, Int64} with 19 stored entries:
⠈⠢⡀⠀⠀⠀⠀⠀⠀⠀
⠀⠀⠈⠢⡀⠀⠀⠀⠀⠀
⠀⠀⠀⠀⠈⠢⡀⠀⠀⠀
⠀⠀⠀⠀⠀⠀⠈⠢⡀⠀
⠀⠀⠀⠀⠀⠀⠀⠀⠈⠢

julia> fock(20, 3)' * a * fock(20, 4)
2.0 + 0.0im

QuPhys.create — Function

create(N::Int)

Bosonic creation operator with Hilbert space cutoff N. This operator acts on a fock state as $\hat{a}^\dagger \ket{n} = \sqrt{n+1} \ket{n+1}$.

Examples

julia> a_d = create(20)
Quantum Object:   type=Operator   dims=[20]   size=(20, 20)   ishermitian=false
20×20 SparseMatrixCSC{ComplexF64, Int64} with 19 stored entries:
⠢⡀⠀⠀⠀⠀⠀⠀⠀⠀
⠀⠈⠢⡀⠀⠀⠀⠀⠀⠀
⠀⠀⠀⠈⠢⡀⠀⠀⠀⠀
⠀⠀⠀⠀⠀⠈⠢⡀⠀⠀
⠀⠀⠀⠀⠀⠀⠀⠈⠢⡀

julia> fock(20, 4)' * a_d * fock(20, 3)
2.0 + 0.0im

QuPhys.sigmap — Function

sigmap()

Pauli ladder operator $\hat{\sigma}_+ = \hat{\sigma}_x + i \hat{\sigma}_y$.

QuPhys.sigmam — Function

sigmam()

Pauli ladder operator $\hat{\sigma}_- = \hat{\sigma}_x - i \hat{\sigma}_y$.

QuPhys.sigmax — Function

sigmax()

Pauli operator $\hat{\sigma}_x = \hat{\sigma}_- + \hat{\sigma}_+$.

QuPhys.sigmay — Function

sigmay()

Pauli operator $\hat{\sigma}_y = i \left( \hat{\sigma}_- - \hat{\sigma}_+ \right)$.

QuPhys.sigmaz — Function

sigmaz()

Pauli operator $\hat{\sigma}_z = \comm{\hat{\sigma}_+}{\hat{\sigma}_-}$.

QuPhys.eye — Function

eye(N::Int; type=OperatorQuantumObject, dims=[N])

Identity operator $\hat{\mathbb{1}}$ with Hilbert dimension N.

QuPhys.fock — Function

fock(N::Int, pos::Int; dims::Vector{Int}=[N], sparse::Bool=false)

Generates a fock state $\ket{\psi}$ of dimension N. It is also possible to specify the list of dimensions dims if different subsystems are present.

QuPhys.basis — Function

basis(N::Int, pos::Int; dims::Vector{Int}=[N])

Generates a fock state like fock.

QuPhys.coherent — Function

coherent(N::Real, α::T)

Generates a coherent state $\ket{\alpha}$, which is defined as an eigenvector of the bosonic annihilation operator $\hat{a} \ket{\alpha} = \alpha \ket{\alpha}$.

QuPhys.rand_dm — Function

rand_dm(N::Integer; kwargs...)

Generates a random density matrix $\hat{\rho}$, with the property to be positive definite, and that $\Tr \left[ \hat{\rho} \right] = 1$.

QuPhys.projection — Function

projection(N::Int, i::Int, j::Int)

Generates the projection operator $\hat{O} = \dyad{i}{j}$ with Hilbert space dimension N.

QuPhys.sinm — Function

sinm(O::QuantumObject)

Generates the sine of the operator O, defined as

$\sin \left( \hat{O} \right) = \frac{e^{i \hat{O}} - e^{-i \hat{O}}}{2 i}$

QuPhys.cosm — Function

cosm(O::QuantumObject)

Generates the cosine of the operator O, defined as

$\cos \left( \hat{O} \right) = \frac{e^{i \hat{O}} + e^{-i \hat{O}}}{2}$

Time evolution

QuPhys.sesolveProblem — Function

sesolveProblem(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector;
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5()
    e_ops::AbstractVector=[],
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    kwargs...)

Generates the ODEProblem for the Schrödinger time evolution of a quantum system.

Arguments

H::QuantumObject: The Hamiltonian of the system.

ψ0::QuantumObject: The initial state of the system.

t_l::AbstractVector: The time list of the evolution.

alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm: The algorithm used for the time evolution.

e_ops::AbstractVector: The list of operators to be evaluated during the evolution.

H_t::Union{Nothing,Function}: The time-dependent Hamiltonian of the system. If nothing, the Hamiltonian is time-independent.

params::Dict{Symbol, Any}: The parameters of the system.

progress::Bool: Whether to show the progress bar.

kwargs...: The keyword arguments passed to the ODEProblem constructor.

Returns

prob: The ODEProblem for the Schrödinger time evolution of the system.

QuPhys.mesolveProblem — Function

mesolveProblem(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::AbstractVector=[];
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::AbstractVector=[],
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    kwargs...)

Generates the ODEProblem for the master equation time evolution of an open quantum system.

Arguments

H::QuantumObject: The Hamiltonian or the Liouvillian of the system.

ψ0::QuantumObject: The initial state of the system.

t_l::AbstractVector: The time list of the evolution.

c_ops::AbstractVector=[]: The list of the collapse operators.

alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(): The algorithm used for the time evolution.

e_ops::AbstractVector=[]: The list of the operators for which the expectation values are calculated.

H_t::Union{Nothing,Function}=nothing: The time-dependent Hamiltonian or Liouvillian.

params::Dict{Symbol, Any}=Dict{Symbol, Any}(): The parameters of the time evolution.

progress::Bool=true: Whether to show the progress bar.

kwargs...: The keyword arguments for the ODEProblem.

Returns

prob::ODEProblem: The ODEProblem for the master equation time evolution.

QuPhys.mcsolveProblem — Function

mcsolveProblem(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::AbstractVector;
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::AbstractVector=[],
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    jump_interp_pts::Integer=-1,
    kwargs...)

Generates the ODEProblem for a single trajectory of the Monte Carlo wave function time evolution of an open quantum system.

Arguments

H::QuantumObject: Hamiltonian of the system.

ψ0::QuantumObject: Initial state of the system.

t_l::AbstractVector: List of times at which to save the state of the system.

c_ops::AbstractVector: List of collapse operators.

alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm: Algorithm to use for the time evolution.

e_ops::AbstractVector: List of operators for which to calculate expectation values.

H_t::Union{Nothing,Function}: Time-dependent part of the Hamiltonian.

params::Dict{Symbol, Any}: Dictionary of parameters to pass to the solver.

jump_interp_pts::Integer: Number of points to use for interpolation of the jump times.

kwargs...: Additional keyword arguments to pass to the solver.

Returns

prob::ODEProblem: The ODEProblem for the Monte Carlo wave function time evolution.

Notes

When jump_interp_pts is set to -1, a DiscreteCallback is used to detect the jump times. When jump_interp_pts is set to a positive integer, a ContinuousCallback is used to detect the jump times.

QuPhys.mcsolveEnsembleProblem — Function

mcsolveEnsembleProblem(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::AbstractVector;
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::AbstractVector=[],
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    n_traj::Integer=1,
    jump_interp_pts::Integer=-1,
    prob_func::Function=_mcsolve_prob_func,
    output_func::Function=_mcsolve_output_func,
    kwargs...)

Generates the ODEProblem for an ensemble of trajectories of the Monte Carlo wave function time evolution of an open quantum system.

Arguments

H::QuantumObject: Hamiltonian of the system.

ψ0::QuantumObject: Initial state of the system.

t_l::AbstractVector: List of times at which to save the state of the system.

c_ops::AbstractVector: List of collapse operators.

alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm: Algorithm to use for the time evolution.

e_ops::AbstractVector: List of operators for which to calculate expectation values.

H_t::Union{Nothing,Function}: Time-dependent part of the Hamiltonian.

params::Dict{Symbol, Any}: Dictionary of parameters to pass to the solver.

progress::Bool: Whether to show a progress bar.

n_traj::Integer: Number of trajectories to use.

jump_interp_pts::Integer: Number of points to use for interpolation of the jump times.

prob_func::Function: Function to use for generating the ODEProblem.

output_func::Function: Function to use for generating the output of a single trajectory.

kwargs...: Additional keyword arguments to pass to the solver.

Returns

prob::EnsembleProblem with ODEProblem: The Ensemble ODEProblem for the Monte Carlo

wave function time evolution.

Notes

When jump_interp_pts is set to -1, a DiscreteCallback is used to detect the jump times. When jump_interp_pts is set to a positive integer, a ContinuousCallback is used to detect the jump times.

QuPhys.sesolve — Function

sesolve(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector;
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::AbstractVector=[],
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    kwargs...)

Time evolution of a closed quantum system using the Schrödinger equation.

Arguments

H::QuantumObject: Hamiltonian of the system.

ψ0::QuantumObject: Initial state of the system.

t_l::AbstractVector: List of times at which to save the state of the system.

alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm: Algorithm to use for the time evolution.

e_ops::AbstractVector: List of operators for which to calculate expectation values.

H_t::Union{Nothing,Function}: Time-dependent part of the Hamiltonian.

params::Dict{Symbol, Any}: Dictionary of parameters to pass to the solver.

progress::Bool: Whether to show a progress bar.

kwargs...: Additional keyword arguments to pass to the solver.

Returns

sol::TimeEvolutionSol: The solution of the time evolution.

QuPhys.mesolve — Function

mesolve(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::AbstractVector=[];
    alg::OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::AbstractVector=[],
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    kwargs...)

Time evolution of an open quantum system using master equation.

Arguments

H::QuantumObject: Hamiltonian of Liouvillian of the system.

ψ0::QuantumObject: Initial state of the system.

t_l::AbstractVector: List of times at which to save the state of the system.

c_ops::AbstractVector: List of collapse operators.

alg::OrdinaryDiffEqAlgorithm: Algorithm to use for the time evolution.

e_ops::AbstractVector: List of operators for which to calculate expectation values.

H_t::Union{Nothing,Function}: Time-dependent part of the Hamiltonian.

params::Dict{Symbol, Any}: Dictionary of parameters to pass to the solver.

progress::Bool: Whether to show a progress bar.

kwargs...: Additional keyword arguments to pass to the solver.

Returns

sol::TimeEvolutionSol: The solution of the time evolution.

QuPhys.mcsolve — Function

mcsolve(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::AbstractVector;
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::AbstractVector=[],
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    n_traj::Integer=1,
    ensemble_method=EnsembleThreads(),
    jump_interp_pts::Integer=-1,
    kwargs...)

Time evolution of an open quantum system using quantum trajectories.

Arguments

H::QuantumObject: Hamiltonian of the system.

ψ0::QuantumObject: Initial state of the system.

t_l::AbstractVector: List of times at which to save the state of the system.

c_ops::AbstractVector: List of collapse operators.

alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm: Algorithm to use for the time evolution.

e_ops::AbstractVector: List of operators for which to calculate expectation values.

H_t::Union{Nothing,Function}: Time-dependent part of the Hamiltonian.

params::Dict{Symbol, Any}: Dictionary of parameters to pass to the solver.

progress::Bool: Whether to show a progress bar.

n_traj::Integer: Number of trajectories to use.

ensemble_method: Ensemble method to use.

jump_interp_pts::Integer: Number of points to use for interpolation of jump times.

kwargs...: Additional keyword arguments to pass to the solver.

Returns

sol::TimeEvolutionMCSol: The solution of the time evolution.

Notes

ensemble_method can be one of EnsembleThreads(), EnsembleSerial(), EnsembleDistributed(). When jump_interp_pts is set to -1, a DiscreteCallback is used to detect the jump times. When jump_interp_pts is set to a positive integer, a ContinuousCallback is used to detect the jump times.

QuPhys.dfd_mesolve — Function

function dfd_mesolve(H::Function, ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::Function, maxdims::AbstractVector;
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::Union{Nothing, Function}=nothing, 
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    tol_list::Vector{<:Number}=Float64[],
    kwargs...)

Time evolution of an open quantum system using master equation, dynamically changing the dimension of the Hilbert subspaces.

QuPhys.dsf_mesolve — Function

function dsf_mesolve(H::Function,
    ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::Function,
    op_list::AbstractVector,
    α0_l::Vector{<:Number}=zeros(length(op_list));
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::Union{Nothing, Function}=nothing,
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    δα_list::Vector{<:Number}=Float64[],
    kwargs...)

Time evolution of an open quantum system using master equation and the Dynamical Shifted Fock algorithm.

QuPhys.dsf_mcsolve — Function

function dsf_mcsolve(H::Function,
    ψ0::QuantumObject,
    t_l::AbstractVector, c_ops::Function,
    op_list::AbstractVector,
    α0_l::Vector{<:Number}=zeros(length(op_list));
    alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm=Tsit5(),
    e_ops::Union{Nothing, Function}=nothing,
    H_t::Union{Nothing,Function}=nothing,
    params::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    progress::Bool=true,
    δα_list::Vector{<:Real}=Float64[],
    n_traj::Integer=1,
    ensemble_method=EnsembleThreads(),
    jump_interp_pts::Integer=10,
    krylov_dim::Integer=cld(prod(ψ0.dims), 4),
    kwargs...)

Time evolution of a quantum system using the Monte Carlo wave function method and the Dynamical Shifted Fock algorithm.

QuPhys.liouvillian — Function

liouvillian(H::QuantumObject, c_ops::AbstractVector)

Construct the Liouvillian superoperator for a system Hamiltonian and a set of collapse operators: $\mathcal{L} \cdot = -i[\hat{H}, \cdot] + \sum_i \mathcal{D}[\hat{O}_i] \cdot$, where $\mathcal{D}[\hat{O}_i] \cdot = \hat{O}_i \cdot \hat{O}_i^\dagger - \frac{1}{2} \hat{O}_i^\dagger \hat{O}_i \cdot - \frac{1}{2} \cdot \hat{O}_i^\dagger \hat{O}_i$.

QuPhys.liouvillian_generalized — Function

liouvillian_generalized(H::QuantumObject, fields::Vector, 
κ_list::Vector, ω_list::Vector, T_list::Vector; N_trunc::Int=size(H,1), tol::Float64=0.0)

Constructs the generalized Liouvillian for a system coupled to a bath of harmonic oscillators.

See, e.g., Settineri, Alessio, et al. "Dissipation and thermal noise in hybrid quantum systems in the ultrastrong-coupling regime." Physical Review A 98.5 (2018): 053834.

QuPhys.steadystate_floquet — Function

steadystate_floquet(H_0::QuantumObject,
    c_ops::Vector, H_p::QuantumObject,
    H_m::QuantumObject,
    ω::Real; n_max::Int=4, lf_solver::LSolver=LiouvillianDirectSolver(),
    ss_solver::Type{SSSolver}=SteadyStateDirectSolver)

Calculates the steady state of a periodically driven system. Here H_0 is the Hamiltonian or the Liouvillian of the undriven system. Considering a monochromatic drive at frequency $\\omega$, we divide it into two parts, H_p and H_m, where H_p oscillates as $e^{i \\omega t}$ and H_m oscillates as $e^{-i \\omega t}$. n_max is the number of iterations used to obtain the effective Liouvillian, lf_solver is the solver used to solve the effective Liouvillian, and ss_solver is the solver used to solve the steady state.

Correlations and Spectrum

QuPhys.correlation_3op_2t — Function

correlation_3op_2t(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector,
    τ_l::AbstractVector,
    A::QuantumObject,
    B::QuantumObject,
    C::QuantumObject,
    c_ops::AbstractVector=[];
    kwargs...)

Returns the two-times correlation function of three operators $\hat{A}$, $\hat{B}$ and $\hat{C}$: $\expval{\hat{A}(t) \hat{B}(t + \tau) \hat{C}(t)}$ for a given initial state $\ket{\psi_0}$.

QuPhys.correlation_2op_2t — Function

correlation_2op_2t(H::QuantumObject,
    ψ0::QuantumObject,
    t_l::AbstractVector,
    τ_l::AbstractVector,
    A::QuantumObject,
    B::QuantumObject,
    c_ops::AbstractVector=[];
    reverse::Bool=false,
    kwargs...)

Returns the two-times correlation function of two operators $\hat{A}$ and $\hat{B}$at different times\expval{\hat{A}(t + \tau) \hat{B}(t)}. Whenreverse=true, the correlation function is calculated as\expval{\hat{A}(t) \hat{B}(t + \tau)}`.

QuPhys.correlation_2op_1t — Function

correlation_2op_1t(H::QuantumObject,
    ψ0::QuantumObject,
    τ_l::AbstractVector,
    A::QuantumObject,
    B::QuantumObject,
    c_ops::AbstractVector=[];
    reverse::Bool=false,
    kwargs...)

Returns the one-time correlation function of two operators $\hat{A}$ and $\hat{B}$ at different times $\expval{\hat{A}(\tau) \hat{B}(0)}$. When $reverse=true$, the correlation function is calculated as $\expval{\hat{A}(0) \hat{B}(\tau)}$.

QuPhys.spectrum — Function

spectrum(H::QuantumObject,
ω_list::AbstractVector,
A::QuantumObject{<:AbstractArray{T2},OperatorQuantumObject},
B::QuantumObject{<:AbstractArray{T3},OperatorQuantumObject},
c_ops::AbstractVector=[];
solver::MySolver=ExponentialSeries(),
    kwargs...)

Returns the emission spectrum $S(\omega) = \int_{-\infty}^\infty \expval{\hat{A}(\tau) \hat{B}(0)} e^{-i \omega \tau} d \tau$.

Eigenvalues and eigenvectors

LinearAlgebra.eigen — Function

LinearAlgebra.eigen(A::QuantumObject; kwargs...)

Calculates the eigenvalues and eigenvectors of the QuantumObject A using the Julia LinearAlgebra package.

julia> a = destroy(5);

julia> H = a + a'
Quantum Object:   type=Operator   dims=[5]   size=(5, 5)   ishermitian=true
5×5 SparseMatrixCSC{ComplexF64, Int64} with 8 stored entries:
     ⋅          1.0+0.0im          ⋅              ⋅          ⋅
 1.0+0.0im          ⋅      1.41421+0.0im          ⋅          ⋅
     ⋅      1.41421+0.0im          ⋅      1.73205+0.0im      ⋅
     ⋅              ⋅      1.73205+0.0im          ⋅      2.0+0.0im
     ⋅              ⋅              ⋅          2.0+0.0im      ⋅

julia> E, U = eigen(H)
Eigen{ComplexF64, Float64, Matrix{ComplexF64}, Vector{Float64}}
values:
5-element Vector{Float64}:
 -2.8569700138728
 -1.3556261799742608
  1.3322676295501878e-15
  1.3556261799742677
  2.8569700138728056
vectors:
5×5 Matrix{ComplexF64}:
  0.106101+0.0im  -0.471249-0.0im  …   0.471249-0.0im  0.106101-0.0im
 -0.303127-0.0im   0.638838+0.0im      0.638838+0.0im  0.303127-0.0im
  0.537348+0.0im  -0.279149-0.0im      0.279149-0.0im  0.537348-0.0im
 -0.638838-0.0im  -0.303127-0.0im     -0.303127-0.0im  0.638838+0.0im
  0.447214+0.0im   0.447214+0.0im     -0.447214-0.0im  0.447214-0.0im

julia> ψ_1 = QuantumObject(U[:,1], dims=H.dims);

julia> expect(H, ψ_1) ≈ E[1]
true

LinearAlgebra.eigvals — Function

LinearAlgebra.eigvals(A::QuantumObject; kwargs...)

Same as eigen(A::QuantumObject; kwargs...) but for only the eigenvalues.

QuPhys.eigsolve — Function

function eigsolve(A::QuantumObject; v0::Union{Nothing,AbstractVector}=nothing, 
    sigma::Union{Nothing, Real}=nothing, k::Integer = min(4, size(A, 1)), 
    krylovdim::Integer = min(10, size(A, 1)), tol::Real = 1e-8, maxiter::Integer = 200,
    solver::Union{Nothing, LinearSolve.SciMLLinearSolveAlgorithm} = nothing, showinfo::Bool=false, kwargs...)

Solve for the eigenvalues and eigenvectors of a matrix A using the Arnoldi method. The keyword arguments are passed to the linear solver.